기능 구현 완료
테스트부탁드립니다

+++ src/main/java/com/khu/cloudcomputing/khuropbox/summarygenerator/repository/GptSummaryRepository.java
@@ -3,5 +3,5 @@ package com.khu.cloudcomputing.khuropbox.summarygenerator.repository;
 import com.khu.cloudcomputing.khuropbox.summarygenerator.entity.GptSummaryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GptSummaryRepository extends JpaRepository<GptSummaryEntity, Long> {
+public interface GptSummaryRepository extends JpaRepository<GptSummaryEntity, Integer> {
 }
